Maintence 1

+++ engine/i_system.h
@@ -46,7 +46,7 @@ boolean        I_StartDisplay(void);
 void            I_EndDisplay(void);
 fixed_t         I_GetTimeFrac(void);
 void            I_GetTime_SaveMS(void);
-unsigned long   I_GetRandomTimeSeed(void);
+uintptr_t   I_GetRandomTimeSeed(void);
 
 // Asynchronous interrupt functions should maintain private queues
 // that are read by the synchronous functions
